+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated dependencies (client-oauth2, graphql, graphql-request, ws) and dev dependencies
 - Added cross-fetch ponyfill
+- Resolve schema issues with transaction assets
 
 ## [0.30.0] - 2020-07-08